Implement option parsing via GOption
We have to activate strict POSIX command line parsing, so that the
option parsers stops at the first non-option. Otherwise we couldn't use
a command structured like this:

dunstcmd --dunst-cmd-option subcmd --subcmd-option

This got introduced in GLib 2.44, so we have to raise the minimum GLib
version. Following the rabbithole, Ubuntu 14.04 doesn't have GLib 2.44,
so we also have to raise the Ubuntu version in TravisCI.